Add a skip for NOT_CRAN
This could fix r-universe build

+++ tests/testit/test-plot.R
@@ -123,7 +123,8 @@ assert(
 
 # should not error when a plot label contains special characters and sanitize=TRUE
 if (requireNamespace('tikzDevice', quietly = TRUE) &&
-    (!is.na(Sys.getenv('CI', NA)) || Sys.getenv('USER') == 'yihui') || Sys.info()[['sysname']] != 'Darwin') {
+    ((!is.na(Sys.getenv('CI', NA)) && !identical(Sys.getenv("NOT_CRAN"), "true")) ||
+     Sys.getenv('USER') == 'yihui') || Sys.info()[['sysname']] != 'Darwin') {
   knit('knit-tikzDevice.Rnw', quiet = TRUE)
   unlink(c('*-tikzDictionary', 'figure', 'knit-tikzDevice.tex'), recursive = TRUE)
 }
